extended search list OBSOLETE list_search_bar revised to include extended search options 	list of buttons that can be used to select a specific field to search in (e.g. title field for text; only search in title) 	special term buttons to indicate whether to combine words or not; and whether and/or columns 	e.g. should a term be present in both fields to return a result
The list_search_bar now includes the link to Add a new instance
changed all list templates to reflect this change

search.py worked with special characters to select specific fields $ and #, these still work but now using the buttons
works as well

work to be done: create an empty search. Search for items that are empty for a given column e.g. search for texts with an empty title.